MET-83: Fixes after review
